% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_noise_fn.R
\name{build_noise_fn}
\alias{build_noise_fn}
\title{Build a standardised noise function}
\usage{
build_noise_fn(type = "gaussian", params = list())
}
\arguments{
\item{type}{Character. Distribution family.}

\item{params}{Named list of parameters for the chosen family.}
}
\value{
A function noise_fn(m) -> numeric vector of length m,
mean 0, variance 1.
}
\description{
Build a standardised noise function
}
