% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A0_hat.R
\name{compute_A0}
\alias{compute_A0}
\title{Compute A_hat_0}
\usage{
compute_A0(T_list, Y_list, h0)
}
\arguments{
\item{T_list}{List of length n. T_list((i)) is the vector of observation
times for subject i.}

\item{Y_list}{List of length n. Y_list((i)) is the vector of observed
values Y_ij for subject i.}

\item{h0}{Numeric scalar. Bandwidth h0 > 0, computed via compute_h0().}
}
\value{
Numeric scalar. The value of A_hat_0.
}
\description{
Compute A_hat_0
}
\examples{
T_list <- list(c(0.1, 0.3, 0.4), c(0.5, 0.6, 0.7, 0.8), c(0.2, 0.4))
Y_list <- list(c(1.2, 0.8, 1.1), c(0.9, 1.3, 1.0, 1.2), c(1.1, 0.7))
compute_A0(T_list, Y_list, h0 = 0.1)
}
