% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_hat.R
\name{compute_B}
\alias{compute_B}
\title{Compute B_hat}
\usage{
compute_B(T_list, h0)
}
\arguments{
\item{T_list}{List of length n. T_list((i)) is the vector of observation
times for subject i.}

\item{h0}{Numeric scalar. Bandwidth h0 > 0, computed via compute_h0().}
}
\value{
Numeric scalar. The value of B_hat, always in (0, 1).
}
\description{
Compute B_hat
}
\examples{
T_list <- list(c(0.1, 0.3, 0.4), c(0.5, 0.6, 0.7, 0.8), c(0.2, 0.4))
compute_B(T_list, h0 = 0.15)
}
