% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cov_from_KL.R
\name{compute_cov_from_KL}
\alias{compute_cov_from_KL}
\title{Compute covariance function from KL specification}
\usage{
compute_cov_from_KL(s, t, eigenfn_list, eigenval_vec)
}
\arguments{
\item{s}{Numeric vector of first set of evaluation points.}

\item{t}{Numeric vector of second set of evaluation points.}

\item{eigenfn_list}{List of K vectorised functions phi_k(t).}

\item{eigenval_vec}{Numeric vector of K positive eigenvalues.}
}
\value{
Numeric matrix of dimension length(s) x length(t) with
C(s_i, t_j) values.
}
\description{
Compute covariance function from KL specification
}
