% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_on_observed_times.R
\name{compute_mu_on_observed_times}
\alias{compute_mu_on_observed_times}
\title{Evaluate mu_hat at the observed times of each subject}
\usage{
compute_mu_on_observed_times(T_list, Y_list, h_mu, kernel, scheme)
}
\arguments{
\item{T_list}{List of length n. T_list((i)) contains the observation times
for subject i.}

\item{Y_list}{List of length n. Y_list((i)) contains the observed values
for subject i.}

\item{h_mu}{Numeric scalar. Bandwidth used for mean estimation.}

\item{kernel}{Character string. Kernel type passed to compute_mu_hat_t().}

\item{scheme}{Character string. Weighting scheme passed to compute_mu_hat_t().}
}
\value{
A list of length n. For each subject i:
\itemize{
\item mu_obs_list((i)) is a numeric vector of length m_i
containing mu_hat(T_ij) for j = 1, ..., m_i.
}
}
\description{
Evaluate mu_hat at the observed times of each subject
}
\examples{
# mu_obs_list <- compute_mu_on_observed_times(
#   T_list = T_list,
#   Y_list = Y_list,
#   h_mu   = h_mu_opt,
#   kernel = "epanechnikov",
#   scheme = "OBS"
# )
}
