% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_covariances.R
\name{compute_raw_covariances}
\alias{compute_raw_covariances}
\title{Compute raw off-diagonal covariance pseudo-observations}
\usage{
compute_raw_covariances(T_list, Y_list, mu_obs_list)
}
\arguments{
\item{T_list}{List of length n. T_list((i)) contains the observation
times for subject i.}

\item{Y_list}{List of length n. Y_list((i)) contains the observed
values for subject i.}

\item{mu_obs_list}{List of length n. mu_obs_list((i)) contains the values
mu_hat(T_ij) for j = 1, ..., m_i.}
}
\value{
A data.frame with one row for each ordered off-diagonal pair (j,l)
within the same subject. Columns:
- id
- j
- l
- t1
- t2
- y1
- y2
- mu1
- mu2
- centered1
- centered2
- raw_cov
}
\description{
Compute raw off-diagonal covariance pseudo-observations
}
\examples{
# mu_obs_list <- compute_mu_on_observed_times(T_list, Y_list, h_mu, kernel, scheme)
# raw_cov_df  <- compute_raw_covariances(T_list, Y_list, mu_obs_list)
}
