% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{compute_residuals}
\alias{compute_residuals}
\title{Compute squared residuals Z_ij = {Y_ij - mu_hat(T_ij)}^2}
\usage{
compute_residuals(T_list, Y_list, h_mu, kernel, scheme)
}
\arguments{
\item{T_list}{List of length n. T_list((i)) is the vector of observation
times for subject i.}

\item{Y_list}{List of length n. Y_list((i)) is the vector of observed
values Y_ij for subject i.}

\item{h_mu}{Numeric scalar. Bandwidth used to estimate mu_hat.}

\item{kernel}{Character string. Kernel type, passed to compute_mu_hat_t().}

\item{scheme}{Character string. Weighting scheme passed to compute_mu_hat_t().
One of "OBS", "SUBJ", "OPTIMAL".}
}
\value{
A list of length n. Z_list((i)) is the vector of squared residuals
Z_ij = {Y_ij - mu_hat(T_ij)}^2 for subject i.
}
\description{
Compute squared residuals Z_ij = {Y_ij - mu_hat(T_ij)}^2
}
