% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sigma_X_from_KL.R
\name{compute_sigma_X_from_KL}
\alias{compute_sigma_X_from_KL}
\title{Compute variance function from KL specification}
\usage{
compute_sigma_X_from_KL(t, eigenfn_list, eigenval_vec)
}
\arguments{
\item{t}{Numeric vector of evaluation points.}

\item{eigenfn_list}{List of K vectorised functions phi_k(t).}

\item{eigenval_vec}{Numeric vector of K positive eigenvalues.}
}
\value{
Numeric vector of same length as t with sigma_X^2(t) values.
}
\description{
Compute variance function from KL specification
}
