% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varsigma2_hat_t.R
\name{compute_varsigma2_hat_t}
\alias{compute_varsigma2_hat_t}
\title{Ridged local linear estimator of varsigma^2(t) at one or more points}
\usage{
compute_varsigma2_hat_t(t_vec, T_list, Z_list, h_sigma, kernel, scheme)
}
\arguments{
\item{t_vec}{Numeric vector. Points at which varsigma^2(t) is estimated.
Can be a scalar for single-point evaluation (backward compatible).}

\item{T_list}{List of length n. T_list((i)) is the vector of observation
times for subject i.}

\item{Z_list}{List of length n. Z_list((i)) is the vector of squared
residuals Z_ij = {Y_ij - mu_hat(T_ij)}^2 for subject i.}

\item{h_sigma}{Numeric scalar. Bandwidth h_sigma > 0.}

\item{kernel}{Character string. Kernel type, passed to kernel_h_sigma().}

\item{scheme}{Character string. Weighting scheme passed to compute_weights().
One of "OBS", "SUBJ", "OPTIMAL".}
}
\value{
Numeric vector of length(t_vec). Returns a scalar if t_vec is scalar.
}
\description{
Ridged local linear estimator of varsigma^2(t) at one or more points
}
