% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_mean_at_times.R
\name{eval_mean_at_times}
\alias{eval_mean_at_times}
\title{Evaluate a mean function at measurement times}
\usage{
eval_mean_at_times(T_list, mu_fn)
}
\arguments{
\item{T_list}{List of n numeric vectors. T_list((i)) contains the m_i
measurement times for subject i, all in (0, 1).}

\item{mu_fn}{Function: numeric vector -> numeric vector of same length.
The user-defined mean function mu(t). Must be vectorised.}
}
\value{
A list of n numeric vectors. mu_values((i)) contains
mu(T_i1), ..., mu(T_im_i).
}
\description{
Evaluate a mean function at measurement times
}
\examples{
# Setup
T_list <- list(c(0.1, 0.2, 0.3), c(0.5, 0.6), c(0.8, 0.9, 0.95, 0.99))
mu_fn  <- function(t) 2 * t^2 * cos(2 * pi * t)

mu_vals <- eval_mean_at_times(T_list, mu_fn)
str(mu_vals)
# List of 3
#  $ : num (1:3) ...
#  $ : num (1:2) ...
#  $ : num (1:4) ...
}
