% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_sigmaX_on_observed_times.R
\name{evaluate_sigmaX_on_observed_times}
\alias{evaluate_sigmaX_on_observed_times}
\title{Evaluate sigma_X(t) at observed times}
\usage{
evaluate_sigmaX_on_observed_times(T_list, res_sigmaX2, rule = 2)
}
\arguments{
\item{T_list}{List of length n. T_list((i)) contains the observation
times for subject i.}

\item{res_sigmaX2}{List. Output of compute_sigmaX2_hat(), containing:
- t_grid
- sigmaX2_hat}

\item{rule}{Integer. Passed to approx() for extrapolation behavior.
Default is 2, meaning constant extrapolation at boundaries.}
}
\value{
A list of length n. For each subject i:
- sigmaX_obs_list((i)) is a numeric vector of length m_i
containing sigmaX_hat(T_ij) = sqrt(sigmaX2_hat(T_ij))
}
\description{
Evaluate sigma_X(t) at observed times
}
\examples{
# res_sigmaX2 <- compute_sigmaX2_hat(res_varsigma, sigma0_hat)
# sigmaX_obs_list <- evaluate_sigmaX_on_observed_times(T_list, res_sigmaX2)
}
