% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_KL_process.R
\name{generate_KL_process}
\alias{generate_KL_process}
\title{Generate centred process via Karhunen-Loève expansion}
\usage{
generate_KL_process(
  T_list,
  eigenfn_list,
  eigenval_vec,
  score_type = "gaussian",
  score_params = list()
)
}
\arguments{
\item{T_list}{List of n numeric vectors of measurement times.}

\item{eigenfn_list}{List of K vectorised functions phi_k(t).}

\item{eigenval_vec}{Numeric vector of K positive eigenvalues.}

\item{score_type}{Character. Distribution family for scores.
Default: "gaussian".}

\item{score_params}{Named list of parameters for the distribution.
Default: list().}
}
\value{
A list with components:
\itemize{
\item Z_list : list of n numeric vectors with Z_i(T_ij) values
\item scores : n x K matrix of scores xi_{ik}
}
}
\description{
Generate centred process via Karhunen-Loève expansion
}
