% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_noise.R
\name{generate_noise}
\alias{generate_noise}
\title{Generate measurement noise for each subject}
\usage{
generate_noise(
  T_list,
  sigma0,
  noise_mean = 0,
  noise_type = "gaussian",
  noise_params = list()
)
}
\arguments{
\item{T_list}{List of n numeric vectors of measurement times.}

\item{sigma0}{Either a single non-negative number (homoscedastic)
or a vectorised function sigma0(t) -> R+
(heteroscedastic).}

\item{noise_mean}{Numeric. Mean of the noise. Must be 0. Declared
explicitly for transparency. Default: 0.}

\item{noise_type}{Character. Distribution family. Default: "gaussian".}

\item{noise_params}{Named list of parameters for the distribution.
Default: list().}
}
\value{
A list of n numeric vectors. noise_list((i)) contains the m_i
noise values for subject i.
}
\description{
Generate measurement noise for each subject
}
