% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{kernel_fun}
\alias{kernel_fun}
\title{Pure kernel function K(u)}
\usage{
kernel_fun(u, kernel = "epanechnikov")
}
\arguments{
\item{u}{Numeric vector. Argument of the kernel.}

\item{kernel}{Character string. One of:
"epanechnikov"  (\emph{) optimal in MSE sense
"biweight"      (}) smoother than Epanechnikov
"triweight"     (\emph{) even smoother
"tricube"       (}) used in LOESS
"triangular"    (\emph{) tent-shaped
"cosine"        (}) cosine-based
"uniform"       (!) not Lipschitz at boundaries
"gaussian"      (!) infinite support}
}
\value{
Numeric vector of the same length as u with values K(u).
}
\description{
Pure kernel function K(u)
}
\examples{
u <- seq(-1.5, 1.5, by = 0.01)
plot(u, kernel_fun(u, "epanechnikov"), type = "l")
lines(u, kernel_fun(u, "biweight"),   col = "red")
lines(u, kernel_fun(u, "gaussian"),   col = "blue")
}
