% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rational_quadratic_correlation.R
\name{rational_quadratic_correlation}
\alias{rational_quadratic_correlation}
\title{Rational Quadratic (Cauchy) correlation function (Lin & Wang 2022)}
\usage{
rational_quadratic_correlation(s, t, theta1, theta2)
}
\arguments{
\item{s}{Numeric vector. First set of time points.}

\item{t}{Numeric vector. Second set of time points.
The function computes the full outer matrix rho(s_i, t_j).}

\item{theta1}{Numeric > 0. Tail decay parameter. Larger values give
faster decay and lighter tails.}

\item{theta2}{Numeric > 0. Scale (range) parameter.}
}
\value{
Numeric matrix of dimension length(s) x length(t).
Entry (i,j) is rho_theta(s_i, t_j).
}
\description{
Rational Quadratic (Cauchy) correlation function (Lin & Wang 2022)
}
\examples{
s <- seq(0, 1, length.out = 5)
t <- seq(0, 1, length.out = 5)
rho <- rational_quadratic_correlation(s, t, theta1 = 1, theta2 = 0.5)
}
