% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho_theta.R
\name{rho_theta}
\alias{rho_theta}
\title{Unified parametric correlation function rho_theta(s, t)}
\usage{
rho_theta(s, t, theta, model)
}
\arguments{
\item{s}{Numeric vector. First set of time points.}

\item{t}{Numeric vector. Second set of time points.}

\item{theta}{Numeric vector of length 2: c(theta1, theta2).}

\item{model}{Character string. One of:
- "matern"
- "power_exponential"
- "rational_quadratic"}
}
\value{
Numeric matrix of dimension length(s) x length(t).
}
\description{
Unified parametric correlation function rho_theta(s, t)
}
\details{
This is a wrapper/dispatcher used in the covariance estimation block.
The actual formulas are delegated to the corresponding functions already
implemented in the data_generation folder.
}
\examples{
s <- c(0.1, 0.3, 0.5)
t <- c(0.2, 0.4)

rho_theta(s, t, theta = c(1, 0.5), model = "power_exponential")
rho_theta(s, t, theta = c(1, 0.5), model = "rational_quadratic")
rho_theta(s, t, theta = c(0.5, 1), model = "matern")
}
