% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_fO_uniform.R
\name{sample_fO_uniform}
\alias{sample_fO_uniform}
\title{Draw n i.i.d. samples from the Uniform density of f_O}
\usage{
sample_fO_uniform(n, delta)
}
\arguments{
\item{n}{Integer >= 1. Number of samples to draw.}

\item{delta}{Numeric in (0, 1). Snippet length. Determines the support
(delta/2, 1 - delta/2).}
}
\value{
Numeric vector of length n with samples O_1, ..., O_n,
all in (delta/2, 1 - delta/2).
}
\description{
Draw n i.i.d. samples from the Uniform density of f_O
}
\examples{
O <- sample_fO_uniform(n = 200, delta = 0.4)
hist(O, breaks = 20, main = "f_O uniform, delta = 0.4")
range(O)   # should be inside (0.2, 0.8)
}
